ctmap: move dummy per-cluster CT map to per_cluster_ctmap_dummy.go
